+++ src/js/Navigation.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '../components/App';
+import Navigation from './Navigation';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<Navigation />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
